+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -927,6 +927,21 @@ class HubTest {
         sut.captureEnvelope(envelope)
         verify(mockClient).captureEnvelope(any(), anyOrNull())
     }
+
+    @Test
+    fun `when captureEnvelope is called, lastEventId is not set`() {
+        val options = SentryOptions().apply {
+            dsn = "https://key@sentry.io/proj"
+            setSerializer(mock())
+        }
+        val sut = Hub(options)
+        val mockClient = mock<ISentryClient>()
+        sut.bindClient(mockClient)
+        whenever(mockClient.captureEnvelope(any(), anyOrNull())).thenReturn(SentryId())
+        val envelope = SentryEnvelope(SentryId(UUID.randomUUID()), null, setOf())
+        sut.captureEnvelope(envelope)
+        assertEquals(SentryId.EMPTY_ID, sut.lastEventId)
+    }
     //endregion
 
     //region startSession tests
@@ -1056,7 +1071,9 @@ class HubTest {
         sut.bindClient(mockClient)
         sut.close()
 
-        sut.captureTransaction(SentryTransaction(SentryTracer(TransactionContext("name", "op"), mock())), null)
+        val sentryTracer = SentryTracer(TransactionContext("name", "op"), sut)
+        sentryTracer.finish()
+        sut.captureTransaction(SentryTransaction(sentryTracer), null)
         verify(mockClient, never()).captureTransaction(any(), any(), any())
     }
 
@@ -1071,13 +1088,44 @@ class HubTest {
         sut.bindClient(mockClient)
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
+        sentryTracer.finish()
         val traceState = sentryTracer.traceState()
-        sut.captureTransaction(SentryTransaction(sentryTracer), traceState)
         verify(mockClient).captureTransaction(any(), eq(traceState), any(), eq(null))
     }
 
     @Test
-    fun `when captureTransaction and transaction is not sampled, captureTransaction on the client should be called`() {
+    fun `when captureTransaction is called, lastEventId is not set`() {
+        val options = SentryOptions().apply {
+            dsn = "https://key@sentry.io/proj"
+            setSerializer(mock())
+        }
+        val sut = Hub(options)
+        val mockClient = mock<ISentryClient>()
+        sut.bindClient(mockClient)
+        whenever(mockClient.captureTransaction(anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull())).thenReturn(SentryId())
+
+        val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
+        sentryTracer.finish()
+        assertEquals(SentryId.EMPTY_ID, sut.lastEventId)
+    }
+
+    @Test
+    fun `when captureTransaction and transaction is not finished, captureTransaction on the client should not be called`() {
+        val options = SentryOptions()
+        options.cacheDirPath = file.absolutePath
+        options.dsn = "https://key@sentry.io/proj"
+        options.setSerializer(mock())
+        val sut = Hub(options)
+        val mockClient = mock<ISentryClient>()
+        sut.bindClient(mockClient)
+
+        val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
+        sut.captureTransaction(SentryTransaction(sentryTracer), null)
+        verify(mockClient, never()).captureTransaction(any(), any(), any(), eq(null))
+    }
+
+    @Test
+    fun `when captureTransaction and transaction is not sampled, captureTransaction on the client should not be called`() {
         val options = SentryOptions()
         options.cacheDirPath = file.absolutePath
         options.dsn = "https://key@sentry.io/proj"
@@ -1087,9 +1135,9 @@ class HubTest {
         sut.bindClient(mockClient)
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", false), sut)
+        sentryTracer.finish()
         val traceState = sentryTracer.traceState()
-        sut.captureTransaction(SentryTransaction(sentryTracer), traceState)
-        verify(mockClient, times(0)).captureTransaction(any(), eq(traceState), any(), eq(null))
+        verify(mockClient, never()).captureTransaction(any(), eq(traceState), any(), eq(null))
     }
     //endregion