+++ sentry/src/test/java/io/sentry/MainEventProcessorTest.kt
@@ -7,6 +7,7 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.hints.ApplyScopeData
 import io.sentry.hints.Cached
+import io.sentry.protocol.DebugMeta
 import io.sentry.protocol.SdkVersion
 import io.sentry.protocol.SentryTransaction
 import io.sentry.protocol.User
@@ -32,13 +33,16 @@ class MainEventProcessorTest {
         val getLocalhost = mock<InetAddress>()
         val sentryTracer = SentryTracer(TransactionContext("", ""), mock())
 
-        fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment", tags: Map<String, String> = emptyMap(), sendDefaultPii: Boolean? = null, serverName: String? = "server", host: String? = null, resolveHostDelay: Long? = null, hostnameCacheDuration: Long = 10): MainEventProcessor {
+        fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment", tags: Map<String, String> = emptyMap(), sendDefaultPii: Boolean? = null, serverName: String? = "server", host: String? = null, resolveHostDelay: Long? = null, hostnameCacheDuration: Long = 10, proguardUuid: String? = null): MainEventProcessor {
             sentryOptions.isAttachThreads = attachThreads
             sentryOptions.isAttachStacktrace = attachStackTrace
             sentryOptions.environment = environment
             sentryOptions.serverName = serverName
             if (sendDefaultPii != null) {
                 sentryOptions.isSendDefaultPii = sendDefaultPii
+            }
+            if (proguardUuid != null) {
+                sentryOptions.proguardUuid = proguardUuid
             }
             tags.forEach { sentryOptions.setTag(it.key, it.value) }
             whenever(getLocalhost.canonicalHostName).thenAnswer {
@@ -402,6 +406,37 @@ class MainEventProcessorTest {
         event = sut.process(event, CustomCachedApplyScopeDataHint())
 
         assertNull(event.threads)
+    }
+
+    @Test
+    fun `when event does not have debug meta and proguard uuids are set, attaches debug information`() {
+        val sut = fixture.getSut(proguardUuid = "id1")
+
+        var event = SentryEvent()
+        event = sut.process(event, null)
+
+        assertNotNull(event.debugMeta) {
+            assertNotNull(it.images) { images ->
+                assertEquals("id1", images[0].uuid)
+                assertEquals("proguard", images[0].type)
+            }
+        }
+    }
+
+    @Test
+    fun `when event has debug meta and proguard uuids are set, attaches debug information`() {
+        val sut = fixture.getSut(proguardUuid = "id1")
+
+        var event = SentryEvent()
+        event.debugMeta = DebugMeta()
+        event = sut.process(event, null)
+
+        assertNotNull(event.debugMeta) {
+            assertNotNull(it.images) { images ->
+                assertEquals("id1", images[0].uuid)
+                assertEquals("proguard", images[0].type)
+            }
+        }
     }
 
     private fun generateCrashedEvent(crashedThread: Thread = Thread.currentThread()) = SentryEvent().apply {