+++ CHANGELOG.md
@@ -2,11 +2,34 @@
 
 ## Unreleased
 
+* Feat: Attach Java vendor and version to events and transactions (#1703)
+
+## 5.2.0
+
+* Feat: Allow setting proguard via Options and/or external resources (#1728)
+* Feat: Add breadcrumbs for the Apollo integration (#1726)
+* Fix: Don't set lastEventId for transactions (#1727)
+* Fix: ActivityLifecycleIntegration#appStartSpan memory leak (#1732)
+
+## 5.2.0-beta.3
+
+* Fix: Check at runtime if AndroidX.Core is available (#1718)
+* Feat: Add "data" to spans (#1717)
+* Fix: Should not capture unfinished transaction (#1719)
+
+## 5.2.0-beta.2
+
+* Bump AGP to 7.0.2 (#1650)
+* Fix: drop spans in BeforeSpanCallback. (#1713)
+
+## 5.2.0-beta.1
+
 * Feat: Add tracestate HTTP header support (#1683)
 * Feat: Add option to filter which origins receive tracing headers (#1698)
 * Feat: Include unfinished spans in transaction (#1699)
 * Fix: Move tags from transaction.contexts.trace.tags to transaction.tags (#1700)
-* Feat: Attach Java vendor and version to events and transactions (#1703)
+* Feat: Add static helpers for creating breadcrumbs (#1702)
+* Feat: Performance support for Android Apollo (#1705)
 
 Breaking changes: